--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -13,7 +13,7 @@ Suggests:
     plm,
     tibble
 Encoding: UTF-8
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.3
 Config/testthat/edition: 3
 LinkingTo: 
     Rcpp,