--- a/R/BMisc.R
+++ b/R/BMisc.R
@@ -6,8 +6,8 @@
 #' @param data data.frame used in function
 #' @param idname unique id
 #' @param tname time period name
-#' @param return_data.table if TRUE, makeBalancedPanel will 
-#'  return a data.table rather than a data.frame.  Default 
+#' @param return_data.table if TRUE, makeBalancedPanel will
+#'  return a data.table rather than a data.frame.  Default
 #'  is FALSE.
 #' @examples
 #' id <- rep(seq(1,100), each = 2) # individual ids for setting up a two period panel
@@ -19,8 +19,8 @@
 #'
 #' @return data.frame that is a balanced panel
 #' @export
-makeBalancedPanel <- function(data, 
-                              idname, 
+makeBalancedPanel <- function(data,
+                              idname,
                               tname,
                               return_data.table=FALSE) {
   if (!inherits(data,"data.frame")) {
@@ -33,8 +33,8 @@ makeBalancedPanel <- function(data,
   if (!return_data.table) {
     return(as.data.frame(data[, if (.N==nt) .SD, by=idname]))
   } else if (return_data.table) {
-    return(data[, if (.N == nt) .SD, by=idname])  
-  } 
+    return(data[, if (.N == nt) .SD, by=idname])
+  }
 }
 
 
@@ -847,20 +847,20 @@ TorF <- function(cond, use_isTRUE=FALSE) {
 
 #' @title get_group_inner
 #' @description Calculates the group for a particular unit
-#' @param this_df a data.frame, for this function it should be specific to 
+#' @param this_df a data.frame, for this function it should be specific to
 #'  a particular unit
 #' @inheritParams get_group
 #' @keywords internal
 #' @export
 get_group_inner <- function(this_df, tname, treatname) {
   if ( all(this_df[,treatname] == 0) ) return(0)
-  
+
   as.numeric( this_df[ this_df[,treatname] == 1, ][1,tname] )
 }
 
 #' @title get_group
 #' @description A function to calculate a unit's group in a panel data setting
-#'  with a binary treatment and staggered treatment adoption and where 
+#'  with a binary treatment and staggered treatment adoption and where
 #'  there is a column in the data indicating whether or not a unit is treated
 #' @param df the data.frame used in the function
 #' @param idname name of column that holds the unit id
@@ -868,19 +868,19 @@ get_group_inner <- function(this_df, tname, treatname) {
 #' @param treatname name of column with the treatment indicator
 #' @export
 get_group <- function(df, idname, tname, treatname) {
-  group_vec <- df %>% 
-    group_by(state) %>% 
+  group_vec <- df %>%
+    group_by(state) %>%
     group_map(~ rep(get_group_inner(.x, tname, treatname), nrow(.x))) %>%
     unlist()
   group_vec
 }
 
 #' @title get_YiGmin1_inner
-#' @description Calculates a units outcome (or also can be used for a covariate) 
-#'  in the period right before it becomes treated.  The unit's group must 
-#'  be specified at this point.  This function operates on a data.frame 
-#'  that is already local to a particular unit.  
-#' @param this_df a data.frame, for this function it should be specific to 
+#' @description Calculates a units outcome (or also can be used for a covariate)
+#'  in the period right before it becomes treated.  The unit's group must
+#'  be specified at this point.  This function operates on a data.frame
+#'  that is already local to a particular unit.
+#' @param this_df a data.frame, for this function it should be specific to
 #'  a particular unit
 #' @inheritParams get_YiGmin1
 #' @keywords internal
@@ -896,26 +896,26 @@ get_YiGmin1_inner <- function(this_df, yname, tname, gname) {
 }
 
 #' @title get_YiGmin1
-#' @description A function to calculate outcomes for units in the period 
-#'  right before they become treated (this function can also be used to recover 
+#' @description A function to calculate outcomes for units in the period
+#'  right before they become treated (this function can also be used to recover
 #'  covariates, etc. in the period right before a unit becomes treated).
-#'  For units that do not 
+#'  For units that do not
 #'  participate in the treatment (and therefore have group==0), they are
 #'  assigned their outcome in the last period.
-#' @param yname name of column containing the outcome (or other variable) 
+#' @param yname name of column containing the outcome (or other variable)
 #'  for which to calculate its outcome in the immediate pre-treatment period
 #' @param gname name of column containing the unit's group
 #' @inheritParams get_group
 #' @export
 get_YiGmin1 <- function(df, idname, yname, tname, gname) {
-  YiGmin1_vec <- df %>% 
-    group_by(.data[[idname]]) %>% 
+  YiGmin1_vec <- df %>%
+    group_by(.data[[idname]]) %>%
     group_map(~ rep(get_YiGmin1_inner(.x, yname, tname, gname), nrow(.x))) %>%
     unlist()
   YiGmin1_vec
 }
 
-#' @title get_Yi1_inner 
+#' @title get_Yi1_inner
 #' @description Calculates a units outcome in the first time period.
 #'  This function operates on a data.frame that is already local to a particular
 #'  unit.
@@ -930,20 +930,20 @@ get_Yi1_inner <- function(this_df, yname, tname, gname) {
 }
 
 #' @title get_Yi1
-#' @description A function to calculate outcomes for units in the first time 
-#'  period that is available in a panel data setting (this function can also 
+#' @description A function to calculate outcomes for units in the first time
+#'  period that is available in a panel data setting (this function can also
 #'  be used to recover covariates, etc. in the first period).
 #' @inheritParams get_YiGmin1
 #' @export
 get_Yi1 <- function(df, idname, yname, tname, gname) {
-  YiGmin1_vec <- df %>% 
-    group_by(.data[[idname]]) %>% 
+  YiGmin1_vec <- df %>%
+    group_by(.data[[idname]]) %>%
     group_map(~ rep(get_Yi1_inner(.x, yname, tname, gname), nrow(.x))) %>%
     unlist()
   YiGmin1_vec
 }
 
-#' @title get_Yibar_inner 
+#' @title get_Yibar_inner
 #' @description Calculates a units average outcome across all periods.
 #'  This function operates on a data.frame that is already local to a particular
 #'  unit.
@@ -956,15 +956,42 @@ get_Yibar_inner <- function(this_df, yname) {
 }
 
 #' @title get_Yibar
-#' @description A function to calculate the average outcome across all time 
-#' periods separately for each unit in a panel data setting (this function can also 
+#' @description A function to calculate the average outcome across all time
+#' periods separately for each unit in a panel data setting (this function can also
 #'  be used to recover covariates, etc.).
 #' @inheritParams get_YiGmin1
 #' @export
 get_Yibar <- function(df, idname, yname) {
-  Yibar_vec <- df %>% 
-    group_by(.data[[idname]]) %>% 
+  Yibar_vec <- df %>%
+    group_by(.data[[idname]]) %>%
     group_map(~ rep(get_Yibar_inner(.x, yname), nrow(.x))) %>%
     unlist()
   Yibar_vec
+}
+
+#' @title get_lagYi
+#' @description A function that calculates lagged outcomes in a panel data setting.
+#'  If the data.frame that is passed in has nxT rows, the resulting vector will
+#'  also have nxT elements with one element for each unit set to be NA
+#' @inheritParams get_Yi1
+#' @param nlags The number of periods to lag.  The default is 1, which computes
+#'  the lag from the previous period.
+#' @export
+get_lagYi <- function(df, idname, yname, tname, nlags=1) {
+    df <- df %>%
+        dplyr::group_by(.data[[idname]]) %>%
+        dplyr::mutate(.lag=dplyr::lag(.data[[yname]],nlags,order_by=.data[[tname]]))
+    df$.lag
+}
+
+#' @title get_first_difference
+#' @description A function that calculates the first difference in a panel data
+#'  setting.  If the data.frame that is passed in has nxT rows, the resulting
+#'  vector will also have nxT elements with one element for each unit set to be
+#'  NA.
+#'  @inheritParams get_lagYi
+#'  @export
+get_first_difference <- function(df, idname, yname, tname) {
+    df$.lag <- get_lagYi(df, idname, yname, tname)
+    df[,yname] - df$.lag
 }