--- a/man/get_Yi1_inner.Rd
+++ b/man/get_Yi1_inner.Rd
@@ -7,10 +7,10 @@
 get_Yi1_inner(this_df, yname, tname, gname)
 }
 \arguments{
-\item{this_df}{a data.frame, for this function it should be specific to 
+\item{this_df}{a data.frame, for this function it should be specific to
 a particular unit}
 
-\item{yname}{name of column containing the outcome (or other variable) 
+\item{yname}{name of column containing the outcome (or other variable)
 for which to calculate its outcome in the immediate pre-treatment period}
 
 \item{tname}{name of column that holds the time period}