--- a/man/get_YiGmin1.Rd
+++ b/man/get_YiGmin1.Rd
@@ -11,7 +11,7 @@ get_YiGmin1(df, idname, yname, tname, gname)
 
 \item{idname}{name of column that holds the unit id}
 
-\item{yname}{name of column containing the outcome (or other variable) 
+\item{yname}{name of column containing the outcome (or other variable)
 for which to calculate its outcome in the immediate pre-treatment period}
 
 \item{tname}{name of column that holds the time period}
@@ -19,10 +19,10 @@ for which to calculate its outcome in the immediate pre-treatment period}
 \item{gname}{name of column containing the unit's group}
 }
 \description{
-A function to calculate outcomes for units in the period 
- right before they become treated (this function can also be used to recover 
+A function to calculate outcomes for units in the period
+ right before they become treated (this function can also be used to recover
  covariates, etc. in the period right before a unit becomes treated).
- For units that do not 
+ For units that do not
  participate in the treatment (and therefore have group==0), they are
  assigned their outcome in the last period.
 }