--- a/man/get_group_inner.Rd
+++ b/man/get_group_inner.Rd
@@ -7,7 +7,7 @@
 get_group_inner(this_df, tname, treatname)
 }
 \arguments{
-\item{this_df}{a data.frame, for this function it should be specific to 
+\item{this_df}{a data.frame, for this function it should be specific to
 a particular unit}
 
 \item{tname}{name of column that holds the time period}