--- a/src/BMisc.cpp~
+++ b/src/BMisc.cpp~
@@ -1,2 +1,0 @@
-#include <Rcpp.h>
-using namespace Rcpp;