… CRAN-SUBMISSION
… NAMESPACE
… NEWS.md
… _pkgdown.yml
… docs/404.html
… docs/authors.html
… docs/index.html
… docs/news/index.html
… docs/pkgdown.yml
… docs/reference/BMisc.html
… docs/reference/TorF.html
… docs/reference/addCovToFormla.html
… docs/reference/blockBootSample.html
… docs/reference/checkfun.html
… docs/reference/combineDfs-1.png
… docs/reference/combineDfs.html
… docs/reference/compareBinary.html
… docs/reference/compareSingleBinary.html
… docs/reference/cs2panel.html
… docs/reference/dropCovFromFormla.html
… docs/reference/getListElement.html
… docs/reference/getWeightedDf.html
… docs/reference/getWeightedMean.html
… docs/reference/getWeightedQuantile.html
… docs/reference/getWeightedQuantiles.html
… docs/reference/get_Yi1.html
… docs/reference/get_Yi1_inner.html
… docs/reference/get_YiGmin1.html
… docs/reference/get_YiGmin1_inner.html
… docs/reference/get_group.html